No remap - UI only
Убрано создание по АПИ.
Все тесты в одном.
Всё создание - через UI
